requirements.txt
click>=8.0.0
pillow>=10.0.0
pytest>=7.0.0
pyfiglet>=0.8.post1
termcolor>=2.4.0
